Started Bartok; issue with step
You will also need to make some changes inside the Unity editor before
the code you just wrote will
work. In the Hierarchy pane, select _MainCamera. The attached Bartok
(Script) component is at the
bottom of the inspector. (If you want to move it up, you can click the
gear next to its name and choose
Move Up.) Set the DeckXML field of Bartok (Script) to the DeckXML file
that is in the Resources
folder of the Project pane. (Because the deck remains unchanged [still
13 cards of 4 suits], this is the
same file that was used by Prospector.)
Now select PrefabCard in the _Prefabs folder of the Project pane. You’re
going to need to swap a
new CardBartok (Script) component for the existing CardProspector one.
Next to the Script
variable in the CardProspector (Script) component, click the Target
button and choose CardBartok to
replace it. Alternatively, you could just attach the CardBartok script
to PrefabCard and then remove
the CardProspector (Script) component.
Now when you press Play, you should see a grid of cards just as you saw
in the early stages of
Prospector.